Rephrase primitive parsers as methods rather than classes.
