+++ src/Parsley/Primitives/LabeledParser.cs
@@ -1,28 +1,31 @@
-namespace Parsley.Primitives;
+namespace Parsley;
 
-class LabeledParser<T> : Parser<T>
+partial class Grammar
 {
-    readonly Parser<T> parse;
-    readonly ErrorMessageList errors;
-
-    public LabeledParser(Parser<T> parse, string expectation)
+    /// <summary>
+    /// When parser p consumes any input, Label(p, e) is the same as p.
+    /// When parser p does not consume any input, Label(p, e) is the same
+    /// as p, except any messages are replaced with expectation e.
+    /// </summary>
+    public static Parser<T> Label<T>(Parser<T> parse, string expectation)
     {
-        this.parse = parse;
-        errors = ErrorMessageList.Empty.With(ErrorMessage.Expected(expectation));
-    }
+        var errors = ErrorMessageList.Empty.With(ErrorMessage.Expected(expectation));
 
-    public Reply<T> Parse(Text input)
-    {
-        var start = input.Position;
-        var reply = parse(input);
-        var newPosition = reply.UnparsedInput.Position;
-        if (start == newPosition)
+        return input =>
         {
-            if (reply.Success)
-                reply = new Parsed<T>(reply.Value, reply.UnparsedInput, errors);
-            else
-                reply = new Error<T>(reply.UnparsedInput, errors);
-        }
-        return reply;
+            var start = input.Position;
+            var reply = parse(input);
+            var newPosition = reply.UnparsedInput.Position;
+
+            if (start == newPosition)
+            {
+                if (reply.Success)
+                    reply = new Parsed<T>(reply.Value, reply.UnparsedInput, errors);
+                else
+                    reply = new Error<T>(reply.UnparsedInput, errors);
+            }
+
+            return reply;
+        };
     }
 }
